requirement_pyqt5.txt
PyQt5
pyexiv2
pyyaml
screeninfo
opencv-python-headless